+++ examples/sequences.rs
@@ -62,7 +62,7 @@ fn main() -> anyhow::Result<()> {
     backend.start_processing(&mut graph, resources)?;
 
     // Create a custom output node. Right now, it just acts like a brickwall limiter.
-    let output_node = graph.push_gen(
+    let output_node = graph.push(
         gen(move |ctx, _resources| {
             let out0 = ctx.outputs.get_channel_mut(0);
             let out1 = ctx.outputs.get_channel_mut(1);
@@ -238,7 +238,7 @@ fn sine_tone_graph(
     graph_settings: GraphSettings,
 ) -> anyhow::Result<Graph> {
     let mut g = Graph::new(graph_settings);
-    let sin = g.push_gen(WavetableOscillatorOwned::new(Wavetable::sine()));
+    let sin = g.push(WavetableOscillatorOwned::new(Wavetable::sine()));
     g.connect(constant(freq).to(sin).to_label("freq"))?;
     let env = Envelope {
         points: vec![(amp, attack), (0.0, duration_secs)],
@@ -247,8 +247,8 @@ fn sine_tone_graph(
         ..Default::default()
     };
     let env = env.to_gen();
-    let env = g.push_gen(env);
-    let mult = g.push_gen(Mult);
+    let env = g.push(env);
+    let mult = g.push(Mult);
     g.connect(sin.to(mult))?;
     g.connect(env.to(mult).to_index(1))?;
     g.connect(mult.to_graph_out())?;
@@ -266,7 +266,7 @@ fn add_sine(
     output_node: NodeAddress,
     main_graph: &mut Graph,
 ) -> anyhow::Result<()> {
-    let node = main_graph.push_graph(sine_tone_graph(
+    let node = main_graph.push(sine_tone_graph(
         freq,
         attack,
         0.05 * amp,